+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ The format is based on `Keep a Changelog <http://keepachangelog.com/en/1.0.0/>`_
 *unreleased*
 ------------
 * added functionality for a ``row_filter`` callable specified during ``apply``
+* added ``keep_columns`` record rule
 * fixed ``__json__`` export warnings with callables